+++ frontend/actions/src/main/java/com/virtuslab/gitmachete/frontend/actions/traverse/TraverseAction.java
@@ -14,6 +14,10 @@ import com.intellij.openapi.progress.Task;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.ui.MessageDialogBuilder;
 import com.intellij.openapi.ui.Messages;
+import com.virtuslab.gitmachete.frontend.actions.backgroundables.OverrideForkPointBackgroundable;
+import com.virtuslab.gitmachete.frontend.actions.backgroundables.RebaseOnParentBackgroundable;
+import com.virtuslab.gitmachete.frontend.actions.backgroundables.ResetCurrentToRemoteBrackgroundable;
+import com.virtuslab.gitmachete.frontend.actions.backgroundables.SlideOutBackgroundable;
 import git4idea.GitLocalBranch;
 import git4idea.branch.GitBranchUiHandlerImpl;
 import git4idea.branch.GitBranchWorker;
